New: Refactor filter model
