+++ packages/core/app/mirage/routes/dashboard.js
@@ -2,7 +2,7 @@ import Mirage from 'ember-cli-mirage';
 import Response from 'ember-cli-mirage/response';
 import moment from 'moment';
 import RESPONSE_CODES from '../enums/response-codes';
-import { getFilterParams, getQueryAuthor } from 'navi-core/utils/rsql-utils';
+import { filterModel } from 'navi-core/utils/rsql-utils';
 
 const TIMESTAMP_FORMAT = 'YYYY-MM-DD HH:mm:ss';
 
@@ -54,31 +54,7 @@ export default function() {
       let ids = idFilter.split(',');
       dashboardObject = dashboards.find(ids);
     } else if ('filter[dashboards]' in request.queryParams) {
-      try {
-        let filterParameters = getFilterParams(queryFilter);
-        let author = getQueryAuthor(queryFilter);
-        if (filterParameters == null && author == null) {
-          throw 'No search parameters';
-        }
-        dashboardObject = dashboards.all().filter(function(dashboard) {
-          // Author can be optional, ie., not included in the query, but filterparameters are always included.
-          const matchesFilterParameterIfExists = filterParameters
-            ? filterParameters.some(filterParameter =>
-                JSON.stringify(dashboard[filterParameter[0]]).match(new RegExp(filterParameter[1], 'i'))
-              )
-            : false;
-          const matchesAuthorIfExists = author ? dashboard.author.id.match(new RegExp(author, 'i')) : true;
-          return matchesFilterParameterIfExists && matchesAuthorIfExists;
-        });
-      } catch (error) {
-        dashboardObject = new Mirage.Response(
-          400,
-          { data: {} },
-          {
-            errors: ['InvalidPredicateException: Invalid filter format']
-          }
-        );
-      }
+      dashboardObject = filterModel(dashboards, queryFilter);
     } else {
       dashboardObject = dashboards.all();
     }
